pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "star-pattern-ai"
version = "0.1.0"
description = "AI-powered autonomous discovery of patterns in star fields"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "astropy>=5.3",
    "astroquery>=0.4.6",
    "photutils>=1.8",
    "sep>=1.2",
    "torch>=2.0",
    "torchvision>=0.15",
    "scikit-learn>=1.3",
    "openai>=1.0",
    "anthropic>=0.18",
    "google-generativeai>=0.3",
    "click>=8.1",
    "rich>=13.0",
    "tqdm>=4.65",
    "Pillow>=10.0",
    "requests>=2.31",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x>=12.0"]
ml-extra = ["umap-learn>=0.5", "hdbscan>=0.8"]
local = ["llama-cpp-python>=0.2", "huggingface_hub>=0.20"]
survey = ["astropy-healpix>=1.0"]
timeseries = ["astroquery>=0.4.7"]
report = ["reportlab>=4.0"]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
star-pattern = "star_pattern.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "network: tests requiring network access",
    "gpu: tests requiring GPU",
    "llm: tests requiring LLM API keys",
    "slow: slow-running tests",
]
